repitbot-microservices-v2/services/auth-service/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.0
psycopg2-binary>=2.9
python-jose>=3.3
alembic>=1.12
aio-pika>=9.4
